backend/requirements/base.txt
alabaster==0.7.13
asgiref==3.6.0
asttokens==2.2.1
attrs==22.2.0
Babel==2.11.0
backcall==0.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.0
cryptography==39.0.1
decorator==5.1.1
defusedxml==0.7.1
django==4.1.7
django-environ==0.9.0
django-extensions==3.1.5
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docutils==0.19
executing==1.2.0
factory-boy==3.2.1
Faker==15.3.3
freezegun==1.2.2
idna==3.4
imagesize==1.4.1
iniconfig==2.0.0
ipdb==0.13.9
ipython==8.10.0
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
oauthlib==3.2.2
packaging==23.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
prompt-toolkit==3.0.37
psycopg2==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
pytest==7.2.0
pytest-cov==3.0.0
pytest-django==4.5.1
pytest-mock==3.7.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
snowballstemmer==2.2.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
Sphinx==5.3.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.3
stack-data==0.6.2
toml==0.10.2
traitlets==5.9.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
wcwidth==0.2.6
whitenoise==5.2.0
